[Refactor] Move ParseTool to base module for better code organization
