+++ autoscholar/__init__.py
@@ -1,0 +1,5 @@
+"""
+AutoScholar - A tool for academic paper processing and analysis
+"""
+
+__version__ = "0.1.0"
