+++ autoscholar/base/__init__.py
@@ -1,0 +1,3 @@
+"""
+Base module containing core abstractions and interfaces
+"""
